requirements.txt
discord.py==2.3.2
google-cloud-vision==3.8.0
python-dotenv
deepl
